Break out classes into their own files
* Makes it easier to navigate/read
